requirements.txt
Flask==3.1.0
matplotlib==3.9.2
python-dotenv==1.0.1